+++ packages/core/kind/computed.js
@@ -20,7 +20,7 @@
  * @param   {Object}    cfg  Configuration object mapping transformation functions to property names
  * @returns {Function}       Function that accepts a props object and mutates it to include the
  *                           computed props
- * @public
+ * @private
  */
 const computed = (cfg, optProps) => {
 	const keys = Object.keys(cfg);