+++ packages/core/CHANGELOG.md
@@ -20,6 +20,12 @@ No significant changes.
 
 No significant changes.
 
+## [1.13.3] - 2017-01-16
+
+### Fixed
+
+- `core/kind` and `core/hoc` public documentation to be visible
+
 ## [1.13.2] - 2017-12-14
 
 No significant changes.